requirements/base.txt
Django==1.9.2
Pillow==3.1.1
PyJWT==1.4.0
django-crispy-forms==1.6.0
django-filter==0.12.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
psycopg2==2.6.1
sqlparse==0.1.18
